feat: init refactor
